backend/requirements.txt
Flask>=2.0.1
Flask-JWT-Extended>=4.3.1
Flask-Cors>=3.0.10
Flask-SQLAlchemy>=2.5.1
Werkzeug>=2.0.0
